+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Changed
  - Removed instances of TryEnter from presenters
+ - Adding OccupancyManager to rooms
+ - Using new OccupancyManager features for occupancy state
 
 ## [15.0.1] - 2021-08-23
 ### Changed