+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  - Removed instances of TryEnter from presenters
  - Adding OccupancyManager to rooms
  - Using new OccupancyManager features for occupancy state
+ - Fixed AbstractConnectProRoom to not wake up via meeting timeout timer if asleep
 
 ## [15.0.1] - 2021-08-23
 ### Changed